Add a status field in entities to determine when they should be saved, updated, or deleted. Add mappers for load & save data. Add Database test.
