--- liquibase.properties
@@ -1,3 +1,0 @@
-defaultSchemaName=habits_model_schema
-databaseChangeLogTable=habits_service_schema.databasechangelog
-databaseChangeLogLockTable=habits_service_schema.databasechangeloglock
