+++ src/main/java/habitsapp/Main.java
@@ -1,25 +1,40 @@
 package habitsapp;
 
-import habitsapp.migration.Migration;
-import habitsapp.out.MenuConsole;
-import habitsapp.repository.DataLoader;
+import habitsapp.data.database.Migration;
+import habitsapp.ui.out.MenuConsole;
+import habitsapp.data.repository.DataLoader;
+
+import java.io.IOException;
+import java.util.Optional;
+import java.util.Properties;
 
 public class Main {
-
-    public static void main(String[] args) {
-        Migration.migrate();
-        DataLoader.load();
-        String testAccountsMsg = """
-         ______________________________________________________________________________________
-         Test accounts:
-            1. Standard user. Name: "user"; Email: "user@mail.ru"; Password: "UserPassword".
-            2. Administrator. Name: "admin"; Email: "admin@mail.ru"; Password: "AdminPassword".
-         ______________________________________________________________________________________
-         """;
-        System.out.println(testAccountsMsg);
-        MenuConsole menuConsole = new MenuConsole();
-        menuConsole.startGuestMenu();
-        DataLoader.release();
+    /**
+     * Invokes migration, data loading, and starts main menu.
+     * Before application is closed, DataLoader saves changes to database.
+     */
+    public static void main(final String[] args) {
+        Optional<Properties> dbProperties = readDatabaseProperties();
+        if (dbProperties.isEmpty()) {
+            return;
+        }
+        Migration.migrate(dbProperties.get());
+        new DataLoader().load();
+        new MenuConsole().startGuestMenu();
+        new DataLoader().release();
     }
-
+    /**
+     * Loads properties file.
+     */
+    private static Optional<Properties> readDatabaseProperties() {
+        Properties dbProperties = new Properties();
+        try {
+            dbProperties.load(Main.class.getClassLoader().getResourceAsStream("application.properties"));
+        }
+        catch (IOException ex) {
+            System.out.println("Database connection error");
+            return Optional.empty();
+        }
+        return Optional.of(dbProperties);
+    }
 }
