+++ src/main/java/habitsapp/data/models/EntityStatus.java
@@ -1,0 +1,8 @@
+package habitsapp.data.models;
+
+public enum EntityStatus {
+    STABLE,
+    CREATED,
+    UPDATED,
+    DELETED
+}
